Error Inappropriately Thrown
User searches for specific text and gets a number of matches back that exceeds or coincides with the `rowsPerPage` only when prior to searching is already on a page that exceeds the number of pages that would be returned after search results. So if they are on page 3, search for something that returns 2 pages, the error will be thrown.